+++ src/main/java/woowacourse/auth/config/AuthenticationPrincipalConfig.java
@@ -1,7 +1,7 @@
 package woowacourse.auth.config;
 
-import woowacourse.auth.ui.AuthenticationPrincipalArgumentResolver;
-import woowacourse.auth.application.AuthService;
+import woowacourse.auth.controller.AuthenticationPrincipalArgumentResolver;
+import woowacourse.auth.service.AuthService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
