+++ src/main/java/woowacourse/auth/controller/AuthController.java
@@ -1,0 +1,4 @@
+package woowacourse.auth.controller;
+
+public class AuthController {
+}
