--- src/main/java/woowacourse/auth/ui/AuthController.java
@@ -1,4 +1,0 @@
-package woowacourse.auth.ui;
-
-public class AuthController {
-}
