+++ src/main/java/woowacourse/shoppingcart/controller/CustomerController.java
@@ -1,0 +1,4 @@
+package woowacourse.shoppingcart.controller;
+
+public class CustomerController {
+}
