--- src/main/java/woowacourse/shoppingcart/ui/CustomerController.java
@@ -1,4 +1,0 @@
-package woowacourse.shoppingcart.ui;
-
-public class CustomerController {
-}
